split out tab logic
