+++ lifu/simulated_interface.py
@@ -1,793 +1,23 @@
-"""In-memory fake of ``openlifu_sdk.io.LIFUInterface`` for ``--simulate``.
+"""Compatibility shim.
 
-Drives the real :class:`LIFUConnector` end-to-end without any USB I/O.
-The only seam used is :meth:`LIFUConnector._make_interface`, so all of
-the connector's state machine, retry/poll/log code paths run against
-the fake exactly as they do against real hardware. Telemetry frames
-emitted during sonication match the format consumed by
-``LIFUConnector.parse_status_string``.
-
-Thermal model
--------------
-Per-module TX temperature integrates ``dT/dt = k * V^2 * duty`` while
-sonicating and decays toward 25 deg C otherwise (Newton's law,
-``tau = 600 s``). ``k`` is calibrated so that 45 V at 25 % duty rises
-50 deg C over 10 minutes.
+The simulated LIFU device interface has moved into the SDK at
+``openlifu_sdk.ui.simulated_interface``. New code should import from there.
+This module is preserved only so that existing imports of
+``lifu.simulated_interface`` keep working.
 """
 
 from __future__ import annotations
 
-import asyncio
-import json
-import logging
-import math
-import random
-import time
-from typing import List, Optional
-
-from PyQt6.QtCore import QObject, QTimer, pyqtSignal
-
-from openlifu_sdk.io.signal import OWSignal
-from openlifu_sdk.io import LIFUInterfaceStatus
-
-logger = logging.getLogger(__name__)
-
-# k chosen so 45 V * 0.25 duty * 600 s -> 50 deg C
-TX_HEATING_K = 50.0 / (45.0 * 45.0 * 0.25 * 600.0)
-# Newton's-law cooling time constant (seconds). 600 s ~ 10 min half-life-ish.
-TX_COOLING_TAU_S = 600.0
-TX_AMBIENT_C = 25.0
-TX_TEMP_NOISE_SIGMA = 0.2
-HV_TEMP_NOISE_SIGMA = 0.1
-HV_VMON_NOISE_SIGMA = 0.05
-
-# Auto-connect delay after start_monitoring() (seconds). Set to 0 so
-# the simulator reports HV + TX as already connected when the QML
-# bindings first evaluate; this avoids transient "Cannot read property
-# of null" warnings from QML expressions that touch device state during
-# launch and lets pages like Operator build their initial plot URL on first
-# refresh (used by the docs/_capture_screenshots.py helper).
-AUTO_CONNECT_DELAY_S = 0.0
-# How often the run engine emits a temperature heartbeat STATUS frame.
-HEARTBEAT_INTERVAL_MS = 1000
-
-
-def _gauss(sigma: float) -> float:
-    return random.gauss(0.0, sigma)
-
-
-def _format_status_frame(pt_curr: int, pt_total: int,
-                         p_curr: int, p_total: int,
-                         temp_tx: float, temp_amb: float,
-                         status: str = "RUNNING",
-                         mode: str = "SEQUENCE") -> str:
-    """Match the format consumed by ``parse_status_string``."""
-    return (
-        f"STATUS:{status},MODE:{mode},"
-        f"PULSE_TRAIN:[{pt_curr}/{pt_total}],"
-        f"PULSE:[{p_curr}/{p_total}],"
-        f"TEMP_TX:{temp_tx:.2f},"
-        f"TEMP_AMBIENT:{temp_amb:.2f}"
-    )
-
-
-# =============================================================================
-# Per-module thermal model
-# =============================================================================
-
-class _ModuleThermal:
-    """Tracks TX module temperature with simple heating / cooling."""
-
-    def __init__(self, module_idx: int):
-        self.module = module_idx
-        self.temp_c = TX_AMBIENT_C
-        # Per-module +/- 5 % variation in heating coefficient so modules
-        # diverge during a long run.
-        self.k_scale = 1.0 + (random.random() - 0.5) * 0.10
-        self._last_update = time.monotonic()
-
-    def heat_step(self, voltage: float, duty: float, dt_s: float):
-        if dt_s <= 0:
-            return
-        rise = TX_HEATING_K * self.k_scale * voltage * voltage * duty * dt_s
-        self.temp_c += rise
-        self._last_update = time.monotonic()
-
-    def cool_step(self):
-        now = time.monotonic()
-        dt = now - self._last_update
-        self._last_update = now
-        if dt <= 0:
-            return
-        # Exact Newton's-law step (more stable than Euler for big dt).
-        self.temp_c = TX_AMBIENT_C + (self.temp_c - TX_AMBIENT_C) * math.exp(-dt / TX_COOLING_TAU_S)
-
-    def read_temp(self) -> float:
-        self.cool_step()
-        return self.temp_c + _gauss(TX_TEMP_NOISE_SIGMA)
-
-    def read_ambient(self) -> float:
-        return TX_AMBIENT_C + _gauss(0.05)
-
-
-# =============================================================================
-# Simulated TX device
-# =============================================================================
-
-class SimulatedTxDevice:
-    """Implements every attribute / method ``LIFUConnector`` calls on
-    ``interface.txdevice``. Owns per-module thermal state and serves as
-    the emitter for unsolicited STATUS frames during sonication.
-    """
-
-    def __init__(self, num_modules: int = 1):
-        self.num_modules = max(1, int(num_modules))
-        self.signal_connected = OWSignal()
-        self.signal_disconnected = OWSignal()
-        self.signal_data_received = OWSignal()
-        self.signal_error = OWSignal()
-
-        self._connected = False
-        self._async_mode = False
-        self._modules = [_ModuleThermal(i) for i in range(self.num_modules)]
-        # Per-module user_config dicts (with module.sensitivity table).
-        self._user_configs = [self._default_user_config(i) for i in range(self.num_modules)]
-        # Last-applied sequence (kept so set_trigger / get_trigger_json round-trip).
-        self._sequence = {
-            "pulse_interval": 0.1,
-            "pulse_count": 1,
-            "pulse_train_interval": 0.0,
-            "pulse_train_count": 1,
-        }
-        self._pulse = {"frequency": 400_000.0, "duration": 100e-6, "amplitude": 1.0}
-        self._trigger_running = False
-
-    # ---- helpers --------------------------------------------------------
-
-    def _default_user_config(self, idx: int) -> dict:
-        # Plausible 100 - 1000 kHz sensitivity table, V/MPa-ish.
-        # Slight per-module variation so preset voltage scaling isn't a no-op.
-        return {
-            "sn": "SIMULATED",
-            "hwid": "ABCDEFGH",
-            "freq": 400,
-            "hw_ver": "SIM",
-            "fw_ver": "2.0.5",
-            "sdk_ver": "1.0.7",
-            "updated": "2026-05-12 08:00:41",
-            "module": {
-                "id": "txm_400_sim-400k-01",
-                "name": "TXM 400kHz (S/N SIMULATED-400K-01)",
-                "nx": 8,
-                "ny": 8,
-                "pitch": 5,
-                "frequency": 400000.0,
-                "kerf": 0.3,
-                "crosstalk_frac": 0.12,
-                "crosstalk_dist": 0.00505,
-                "sensitivity": [
-                    [375000,3144],
-                    [380000,3110],
-                    [385000,2823],
-                    [390000,2796],
-                    [395000,2744],
-                    [400000,2720],
-                    [405000,2300],
-                    [410000,2267]
-                ]
-            },
-            "device": {}
-        }
-
-    def is_connected(self) -> bool:
-        return self._connected
-
-    def emit_connected(self, port: str = "SIM:TX"):
-        if self._connected:
-            return
-        self._connected = True
-        self.signal_connected.emit("TX", port)
-
-    def emit_disconnected(self, port: str = "SIM:TX"):
-        if not self._connected:
-            return
-        self._connected = False
-        self.signal_disconnected.emit("TX", port)
-
-    def emit_status_frame(self, pt_curr: int, pt_total: int,
-                          p_curr: int = 0, p_total: int = 0,
-                          status: str = "RUNNING",
-                          mode: str = "SEQUENCE") -> None:
-        # Use module 0 temp as the representative one (matches firmware behavior).
-        temp_tx = self._modules[0].read_temp()
-        temp_amb = self._modules[0].read_ambient()
-        frame = _format_status_frame(
-            pt_curr, pt_total, p_curr, p_total, temp_tx, temp_amb,
-            status=status, mode=mode,
-        )
-        self.signal_data_received.emit("TX", frame)
-
-    # ---- methods called by the connector --------------------------------
-
-    def get_tx_module_count(self) -> int:
-        return self.num_modules
-
-    def get_module_count(self) -> int:
-        return self.num_modules
-
-    def get_temperature(self, module: int = 0) -> float:
-        return self._modules[module].read_temp()
-
-    def get_ambient_temperature(self, module: int = 0) -> float:
-        return self._modules[module].read_ambient()
-
-    def get_version(self, module: int = 0) -> str:
-        return "sim-1.0.7"
-
-    def get_hardware_id(self, module: int = 0, raw_hex: bool = False) -> str:
-        # 32-hex-char ID matching HW_ID_DATA_LENGTH (16 bytes) the connector
-        # base58-encodes for display.
-        return f"{0xA0A1A2A3A4A5A6A7B0B1B2B3B4B5B6B7 + module:032X}"
-
-    def read_config(self, module: int = 0):
-        from openlifu_sdk.io.LIFUUserConfig import LifuUserConfig
-        return LifuUserConfig(json_data=dict(self._user_configs[module]))
-
-    def write_config_json(self, json_str: str, module: int = 0):
-        from openlifu_sdk.io.LIFUUserConfig import LifuUserConfig
-        try:
-            self._user_configs[module] = json.loads(json_str)
-        except Exception:
-            logger.warning("SimulatedTxDevice.write_config_json: invalid json; ignored")
-        return LifuUserConfig(json_data=dict(self._user_configs[module]))
-
-    def _normalize_train_interval(self):
-        """Substitute pulse_train_interval=0 with pulse_count*pulse_interval.
-
-        The firmware/SDK convention is that a zero train interval means
-        "back-to-back": the train period equals pulse_count*pulse_interval.
-        We bake that substitution in once at storage time so every
-        downstream consumer (run engine, status frames, get_trigger_json)
-        sees a consistent non-zero value.
-        """
-        try:
-            ti = float(self._sequence.get("pulse_train_interval", 0.0))
-        except (TypeError, ValueError):
-            ti = 0.0
-        if ti > 0:
-            return
-        try:
-            pi = float(self._sequence.get("pulse_interval", 0.0))
-            pc = int(self._sequence.get("pulse_count", 1))
-        except (TypeError, ValueError):
-            pi, pc = 0.0, 1
-        self._sequence["pulse_train_interval"] = max(1e-3, pc * pi)
-
-    def set_solution(self, pulse=None, delays=None, apodizations=None,
-                     sequence=None, profile_index=1, profile_increment=True,
-                     trigger_mode="sequence"):
-        if pulse:
-            self._pulse = dict(pulse)
-        if sequence:
-            self._sequence = dict(sequence)
-            self._normalize_train_interval()
-        return True
-
-    def set_trigger(self, pulse_interval=None, pulse_count=None,
-                    pulse_train_interval=None, pulse_train_count=None,
-                    trigger_mode="sequence"):
-        if pulse_interval is not None:
-            self._sequence["pulse_interval"] = float(pulse_interval)
-        if pulse_count is not None:
-            self._sequence["pulse_count"] = int(pulse_count)
-        if pulse_train_interval is not None:
-            self._sequence["pulse_train_interval"] = float(pulse_train_interval)
-        if pulse_train_count is not None:
-            self._sequence["pulse_train_count"] = int(pulse_train_count)
-        self._normalize_train_interval()
-        return self.get_trigger_json()
-
-    def get_trigger_json(self) -> dict:
-        return {
-            "TriggerStatus": "RUNNING" if self._trigger_running else "STOPPED",
-            "TriggerMode": "SEQUENCE",
-            **self._sequence,
-        }
-
-    def set_trigger_json(self, data) -> dict:
-        if isinstance(data, dict):
-            for k in ("pulse_interval", "pulse_count",
-                      "pulse_train_interval", "pulse_train_count"):
-                if k in data:
-                    self._sequence[k] = data[k]
-            self._normalize_train_interval()
-        return self.get_trigger_json()
-
-    def async_mode(self, enable: Optional[bool] = None) -> bool:
-        if enable is not None:
-            self._async_mode = bool(enable)
-        return self._async_mode
-
-    def start_trigger(self):
-        self._trigger_running = True
-
-    def stop_trigger(self):
-        self._trigger_running = False
-
-    def set_module_invert(self, invert):
-        return None
-
-    def ping(self, module: int = 0):
-        return True
-
-    def toggle_led(self, module: int = 0):
-        return True
-
-    def echo(self, echo_data: bytes, module: int = 0):
-        return (echo_data, len(echo_data))
-
-    def soft_reset(self, module: Optional[int] = None):
-        return True
-
-    def update_firmware(self, *args, **kwargs):
-        # Verification tests / FW updater aren't in the simulator scope.
-        raise NotImplementedError("Firmware update not supported in simulation mode")
-
-    def close(self):
-        self._connected = False
-
-    async def start_monitoring(self, interval: int = 1):
-        return None
-
-    def stop_monitoring(self):
-        return None
-
-
-# =============================================================================
-# Simulated HV controller
-# =============================================================================
-
-class SimulatedHVController:
-    """Implements every attribute / method ``LIFUConnector`` calls on
-    ``interface.hvcontroller``."""
-
-    def __init__(self):
-        self.signal_connected = OWSignal()
-        self.signal_disconnected = OWSignal()
-        self.signal_data_received = OWSignal()
-        self.signal_error = OWSignal()
-
-        self._connected = False
-        self._hv_on = False
-        self._v12_on = True
-        self._voltage_setpoint = 0.0
-        self._rgb_state = 0
-        self.uart = None  # connector reads this for FW DFU; not used here
-
-    def is_connected(self) -> bool:
-        return self._connected
-
-    def emit_connected(self, port: str = "SIM:CON"):
-        if self._connected:
-            return
-        self._connected = True
-        self.signal_connected.emit("HV", port)
-
-    def emit_disconnected(self, port: str = "SIM:CON"):
-        if not self._connected:
-            return
-        self._connected = False
-        self.signal_disconnected.emit("HV", port)
-
-    # ---- methods --------------------------------------------------------
-
-    def turn_hv_on(self):
-        self._hv_on = True
-        return True
-
-    def turn_hv_off(self):
-        self._hv_on = False
-        return True
-
-    def get_hv_status(self) -> bool:
-        return self._hv_on
-
-    def turn_12v_on(self):
-        self._v12_on = True
-        return True
-
-    def turn_12v_off(self):
-        self._v12_on = False
-        return True
-
-    def get_12v_status(self) -> bool:
-        return self._v12_on
-
-    def get_version(self) -> str:
-        return "sim-1.0.7"
-
-    def get_hardware_id(self, raw_hex: bool = False) -> str:
-        return "C0C1C2C3C4C5C6C7D0D1D2D3D4D5D6D7"
-
-    def get_temperature1(self) -> float:
-        return 30.0 + 0.05 * self._voltage_setpoint + _gauss(HV_TEMP_NOISE_SIGMA)
-
-    def get_temperature2(self) -> float:
-        return 31.0 + 0.05 * self._voltage_setpoint + _gauss(HV_TEMP_NOISE_SIGMA)
-
-    def set_voltage(self, voltage: float) -> bool:
-        self._voltage_setpoint = float(voltage)
-        return True
-
-    def get_voltage(self) -> float:
-        return self._voltage_setpoint if self._hv_on else 0.0
-
-    def get_vmon_values(self) -> List[dict]:
-        """Match the real SDK shape: list of 8 dicts with channel, raw_adc,
-        voltage, and converted_voltage fields. QML reads ``converted_voltage``.
-        """
-        v = self._voltage_setpoint if self._hv_on else 0.0
-        v12 = 12.0 + _gauss(HV_VMON_NOISE_SIGMA) if self._v12_on else 0.0
-        converted = [
-            +v + _gauss(HV_VMON_NOISE_SIGMA),       # HVP1
-            +v + _gauss(HV_VMON_NOISE_SIGMA),       # HVP2
-            -v + _gauss(HV_VMON_NOISE_SIGMA),       # HVM2
-            -v + _gauss(HV_VMON_NOISE_SIGMA),       # HVM1
-            v12,                                     # 12V
-            3.3 + _gauss(0.01),                     # VCA1
-            3.3 + _gauss(0.01),                     # VCB1
-            1.8 + _gauss(0.01),                     # VCC1
-        ]
-        return [
-            {
-                "channel": i,
-                "raw_adc": int(max(0, min(65535, abs(cv) * 1000))),
-                "voltage": round(cv, 3),
-                "converted_voltage": round(cv, 3),
-            }
-            for i, cv in enumerate(converted)
-        ]
-
-    def set_rgb_led(self, state: int):
-        self._rgb_state = int(state)
-        return True
-
-    def get_rgb_led(self) -> int:
-        return self._rgb_state
-
-    def ping(self):
-        return True
-
-    def toggle_led(self):
-        return True
-
-    def echo(self, echo_data: bytes):
-        return (echo_data, len(echo_data))
-
-    def soft_reset(self):
-        return True
-
-    def enter_dfu(self):
-        raise NotImplementedError("DFU not supported in simulation mode")
-
-    def close(self):
-        self._connected = False
-
-    async def start_monitoring(self, interval: int = 1):
-        return None
-
-    def stop_monitoring(self):
-        return None
-
-
-# =============================================================================
-# Run engine - emits STATUS frames during sonication
-# =============================================================================
-
-class _SimulatedRunEngine(QObject):
-    """Drives one sonication run: emits STATUS frames and applies thermal
-    heating to the TX modules at the configured pulse-train cadence.
-
-    The engine lives on the main thread; both timers are QTimers parented
-    to it. ``alive`` flips False when the run finishes, which the
-    connector's ``queryTxTemperature()`` polling sees via
-    ``LIFUInterface.is_running()`` and uses to drive its own RUNNING ->
-    READY transition.
-    """
-
-    finished = pyqtSignal()
-
-    def __init__(self, txdevice: SimulatedTxDevice, hvcontroller: SimulatedHVController,
-                 sequence: dict, pulse: dict, voltage: float,
-                 trigger_mode: str = "sequence", parent=None):
-        super().__init__(parent)
-        self._tx = txdevice
-        self._hv = hvcontroller
-        self._voltage = float(voltage)
-        self._trigger_mode = str(trigger_mode).lower()
-        self._mode_label = {
-            "sequence": "SEQUENCE",
-            "continuous": "CONTINUOUS",
-            "single": "SINGLE",
-        }.get(self._trigger_mode, "SEQUENCE")
-
-        # Effective pulse-train period: when pulse_train_interval is 0
-        # the SDK uses pulse_count * pulse_interval.
-        pulse_interval = float(sequence.get("pulse_interval", 0.1))
-        pulse_count = int(sequence.get("pulse_count", 1))
-        train_interval = float(sequence.get("pulse_train_interval", 0.0))
-        self._pulse_count = pulse_count
-        self._pulse_interval_s = pulse_interval
-        self._train_period_s = train_interval if train_interval > 0 else max(
-            1e-3, pulse_count * pulse_interval
-        )
-        # Trigger-mode shapes the train-count semantics:
-        #   sequence   - run pulse_train_count trains, then STOPPED
-        #   single     - run exactly one train, then STOPPED
-        #   continuous - run forever (PT[1/1] held), only stops on
-        #                explicit stop_sonication() from the host
-        seq_total = max(1, int(sequence.get("pulse_train_count", 1)))
-        if self._trigger_mode == "single":
-            self._train_total = 1
-            self._infinite = False
-        elif self._trigger_mode == "continuous":
-            self._train_total = 1
-            self._infinite = True
-        else:
-            self._train_total = seq_total
-            self._infinite = False
-
-        # Duty for thermal model.
-        pulse_duration_s = float(pulse.get("duration", 0.0))
-        self._duty = (pulse_count * pulse_duration_s) / self._train_period_s if self._train_period_s > 0 else 0.0
-
-        self._train_curr = 0
-        self.alive = True
-
-        self._train_timer = QTimer(self)
-        self._train_timer.setSingleShot(False)
-        self._train_timer.timeout.connect(self._on_train_tick)
-
-        self._heartbeat = QTimer(self)
-        self._heartbeat.setSingleShot(False)
-        self._heartbeat.timeout.connect(self._on_heartbeat)
-
-    def start(self):
-        self._tx.start_trigger()
-        # Apply heating for the very first train period as it elapses;
-        # speed-clamp to avoid pegging the GUI on tiny periods.
-        period_ms = max(20, int(round(self._train_period_s * 1000)))
-        if self._infinite:
-            est_duration = "infinite"
-        else:
-            est_duration = f"{self._train_total * self._train_period_s:.3f}s"
-        logger.info(
-            "[SIMRUN] start mode=%s pulse_count=%d pulse_interval=%.4fs "
-            "train_period=%.4fs (timer=%dms) train_total=%s "
-            "expected_duration=%s",
-            self._mode_label, self._pulse_count, self._pulse_interval_s,
-            self._train_period_s, period_ms,
-            "inf" if self._infinite else str(self._train_total),
-            est_duration,
-        )
-        # Emit an initial RUNNING frame at PT[0/N] so the UI flips into
-        # the running state immediately rather than waiting one full
-        # train period for the first tick. This matters when the train
-        # period is long (multi-second sequences) AND when it's short
-        # but the count is low (so the user otherwise sees nothing
-        # before STOPPED).
-        initial_total = self._train_curr if self._infinite else self._train_total
-        self._tx.emit_status_frame(
-            self._train_curr, max(1, initial_total),
-            status="RUNNING", mode=self._mode_label,
-        )
-        self._train_timer.start(period_ms)
-        self._heartbeat.start(HEARTBEAT_INTERVAL_MS)
-
-    def stop(self):
-        was_alive = self.alive
-        # Mark inactive first so any queued timer ticks become no-ops.
-        self.alive = False
-        self._train_timer.stop()
-        self._heartbeat.stop()
-        self._tx.stop_trigger()
-        # Emit a final STOPPED frame so the connector's STATUS-based
-        # trigger-state machine flips cleanly (especially in continuous
-        # mode where there's no natural completion).
-        if was_alive:
-            if self._infinite:
-                total = self._train_curr if self._train_curr > 0 else 1
-            else:
-                total = self._train_total
-            self._tx.emit_status_frame(
-                self._train_curr, total,
-                status="STOPPED", mode=self._mode_label,
-            )
-
-    def _on_train_tick(self):
-        if not self.alive:
-            return
-        self._train_curr += 1
-        # Apply heating for this train period.
-        for m in self._tx._modules:
-            m.heat_step(self._voltage, self._duty, self._train_period_s)
-        if self._infinite:
-            # Continuous mode: emit PT[curr/curr] so the counter keeps
-            # ticking up forever; only stops on explicit stop_sonication.
-            self._tx.emit_status_frame(
-                self._train_curr, self._train_curr,
-                status="RUNNING", mode=self._mode_label,
-            )
-            return
-        self._tx.emit_status_frame(
-            self._train_curr, self._train_total,
-            status="RUNNING", mode=self._mode_label,
-        )
-        if self._train_curr >= self._train_total:
-            # Mark inactive BEFORE emitting so a queued heartbeat tick
-            # cannot race past us and re-emit a RUNNING frame that
-            # would clobber the state reset on the connector side.
-            self.alive = False
-            self._train_timer.stop()
-            self._heartbeat.stop()
-            # Final STOPPED frame so the connector flips trigger state /
-            # transitions back to READY.
-            self._tx.emit_status_frame(
-                self._train_curr, self._train_total,
-                status="STOPPED", mode=self._mode_label,
-            )
-            self.finished.emit()
-
-    def _on_heartbeat(self):
-        if not self.alive:
-            return
-        # Carry latest progress + temperature between train ticks.
-        if self._infinite:
-            total = self._train_curr if self._train_curr > 0 else 1
-        else:
-            total = self._train_total
-        self._tx.emit_status_frame(
-            self._train_curr, total,
-            status="RUNNING", mode=self._mode_label,
-        )
-
-
-# =============================================================================
-# Simulated LIFUInterface (top-level fake)
-# =============================================================================
-
-class SimulatedLIFUInterface(QObject):
-    """Drop-in fake for :class:`openlifu_sdk.io.LIFUInterface`."""
-
-    def __init__(self, num_modules: int = 1,
-                 voltage_table_selection: Optional[str] = None,
-                 sequence_time_selection: Optional[str] = None,
-                 duty_cycle_selection: Optional[str] = None,
-                 **_unused):
-        super().__init__()
-        self.txdevice = SimulatedTxDevice(num_modules=num_modules)
-        self.hvcontroller = SimulatedHVController()
-        self.status = LIFUInterfaceStatus.STATUS_SYS_OFF
-        self._engine: Optional[_SimulatedRunEngine] = None
-        # Mirror real interface's attributes so any external code that
-        # peeks at them sees something sensible.
-        self.voltage_table_selection = voltage_table_selection
-        self.sequence_time_selection = sequence_time_selection
-        self.duty_cycle_selection = duty_cycle_selection
-        self._last_solution_voltage = 0.0
-        self._last_trigger_mode = "sequence"
-
-    # ---- monitoring lifecycle -------------------------------------------
-
-    async def start_monitoring(self, interval: int = 1):
-        # Auto-connect both devices ~AUTO_CONNECT_DELAY_S after launch
-        # via QTimer so the connect signals are delivered on the GUI
-        # thread (mirroring the real OWSignal -> _Bridge path).
-        delay_ms = int(AUTO_CONNECT_DELAY_S * 1000)
-
-        def _connect():
-            logger.info("SimulatedLIFUInterface: emitting auto-connect for HV + TX")
-            self.hvcontroller.emit_connected()
-            self.txdevice.emit_connected()
-
-        QTimer.singleShot(delay_ms, _connect)
-        return None
-
-    def stop_monitoring(self):
-        return None
-
-    def is_device_connected(self):
-        return (self.txdevice.is_connected(), self.hvcontroller.is_connected())
-
-    # ---- solution / sonication ------------------------------------------
-
-    def set_solution(self, solution, profile_index=1, profile_increment=True,
-                     trigger_mode="sequence", turn_hv_on: bool = False,
-                     wait_for_settle: bool = False,
-                     _allow_unsafe_solution: bool = False):
-        """Skip safety checks; just store the bits the run engine needs."""
-        voltage = float(solution.get("voltage", 0.0))
-        self._last_solution_voltage = voltage
-        self._last_trigger_mode = str(trigger_mode).lower()
-        self.txdevice.set_solution(
-            pulse=solution.get("pulse"),
-            sequence=solution.get("sequence"),
-            trigger_mode=trigger_mode,
-        )
-        # Real LIFUInterface.set_solution pushes the voltage setpoint
-        # down to the HV controller as part of loading the solution.
-        # Mirror that so QML's vmon plots / rail readouts track the
-        # configured value.
-        self.hvcontroller.set_voltage(voltage)
-        self.set_status(LIFUInterfaceStatus.STATUS_READY)
-        if turn_hv_on:
-            self.hvcontroller.turn_hv_on()
-        return True
-
-    def start_sonication(self, async_mode: Optional[bool] = None,
-                         turn_hv_on: bool = True,
-                         wait_for_settle: bool = True) -> bool:
-        if turn_hv_on:
-            self.hvcontroller.turn_hv_on()
-        if wait_for_settle:
-            # Brief settle delay (real device is ~200 ms); not perceptible
-            # but matches the real code path's blocking nature.
-            time.sleep(0.2)
-        # Stop any previous engine before starting a new one (pause/resume
-        # rebuilds the trigger then re-calls start_sonication).
-        if self._engine is not None and self._engine.alive:
-            self._engine.stop()
-        self.txdevice.async_mode(True)
-        self._engine = _SimulatedRunEngine(
-            txdevice=self.txdevice,
-            hvcontroller=self.hvcontroller,
-            sequence=self.txdevice._sequence,
-            pulse=self.txdevice._pulse,
-            voltage=self._last_solution_voltage,
-            trigger_mode=self._last_trigger_mode,
-            parent=self,
-        )
-        self._engine.start()
-        self.set_status(LIFUInterfaceStatus.STATUS_RUNNING)
-        return True
-
-    def stop_sonication(self, turn_hv_off: bool = True,
-                        wait_for_settle: bool = False) -> bool:
-        if self._engine is not None:
-            self._engine.stop()
-            self._engine = None
-        self.txdevice.async_mode(False)
-        if turn_hv_off:
-            self.hvcontroller.turn_hv_off()
-        self.set_status(LIFUInterfaceStatus.STATUS_READY)
-        return True
-
-    def is_running(self) -> bool:
-        return self._engine is not None and self._engine.alive
-
-    # ---- misc -----------------------------------------------------------
-
-    def set_status(self, status: LIFUInterfaceStatus):
-        self.status = status
-
-    def get_status(self) -> LIFUInterfaceStatus:
-        return self.status
-
-    def check_solution(self, solution):  # always passes
-        return None
-
-    def set_module_invert(self, module_invert):
-        self.txdevice.set_module_invert(module_invert)
-
-    def close(self):
-        if self._engine is not None:
-            self._engine.stop()
-            self._engine = None
-        try:
-            self.hvcontroller.close()
-        except Exception:
-            pass
-        try:
-            self.txdevice.close()
-        except Exception:
-            pass
+from openlifu_sdk.ui.simulated_interface import (
+    SimulatedHVController,
+    SimulatedLIFUInterface,
+    SimulatedTxDevice,
+)
+from openlifu_sdk.ui.status_frame import format_status_frame as _format_status_frame
+
+__all__ = [
+    "SimulatedHVController",
+    "SimulatedLIFUInterface",
+    "SimulatedTxDevice",
+    "_format_status_frame",
+]
